requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
google-api-python-client==2.139.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
requests==2.32.3
